+++ .nycrc.json
@@ -5,12 +5,11 @@
   ],
   "exclude": [
     "src/config/*.js",
-    "src/middlewares/*.js",
-    "src/routes/*.js"
+    "src/service/*.js"
   ],
   "check-coverage": true,
-  "branches": 80,
-  "lines": 80,
-  "functions": 80,
-  "statements": 80
+  "branches": 85,
+  "lines": 85,
+  "functions": 85,
+  "statements": 85
 }
