+++ src/routes/health.js
@@ -4,7 +4,7 @@ const express = require('express')
 const route = express.Router()
 const asyncMiddleware = require('../middlewares/asyncMiddleware')
 route.get(
-  '/health',
+  '/',
   asyncMiddleware(async (req, res) => {
     await res.status(200).send('Healthy')
   })
